+++ src/schrodinger/sketcher/menu/abstract_context_menu.h
@@ -36,6 +36,9 @@ class SKETCHER_API AbstractContextMenu : public QMenu
      * 3. immediately hidden when an action is clicked
      * In other words, we expect that the model will not change for the
      * duration of the context menu being shown.
+     *
+     * `primary_atom` is the atom under the cursor at right-click (or null);
+     * see `m_primary_atom` for how it relates to `atoms`.
      */
     virtual void setContextItems(
         const std::unordered_set<const RDKit::Atom*>& atoms,
